Improving Deutsch Algorithm notebook
Improved the markdown cells at the beginning to add professionality and coherence.
Added to do list to finalize the notebook for commit and merge to main branch.
Added a results section at the end for the user to better visualize the results of it, and to see the effect of noisy computation vs ideal computation.